+++ src/types.ts
@@ -1,6 +1,3 @@
-import { Model } from './model.ts';
-import { Store } from './store.ts';
-
 export interface JsonApiDocument<Type extends string = string> {
     data: JsonApiResource<Type> | JsonApiResource<Type>[] | null;
     included?: JsonApiResource[];
@@ -40,17 +37,3 @@ export interface JsonApiRelationshipToMany<Type extends string = string>
 }
 
 export type SchemaCollection = { [Type in string]: JsonApiResource<Type> };
-
-export type ModelMap<Schemas extends SchemaCollection = SchemaCollection> = {
-    [Type in keyof Schemas & string]?: new (
-        data: JsonApiResource<Type>,
-        store: Store<Schemas>,
-    ) => Schemas[Type];
-};
-
-export type ModelForType<
-    Type extends string,
-    Schemas extends SchemaCollection,
-> = Type extends keyof Schemas
-    ? Model<Schemas[Type], Schemas>
-    : Model<JsonApiResource<Type>, Schemas>;